+++ src/plc/mock_plc_facade.c
@@ -4,6 +4,6 @@ int configWritePlcMessage([[maybe_unused]] PlcMessage_t* msg,
                           [[maybe_unused]] PlcMessageType_e msgType,
                           [[maybe_unused]] uint16_t station,
                           [[maybe_unused]] enum TrainId_e trainId,
-                          [[maybe_unused]] int target) {
+                          [[maybe_unused]] uint16_t target) {
   return 0;
 }
